pyproject.toml
[tool.poetry]
name = "fabelio_price_tracker"
version = "1.0.0"
description = "Price tracking for fabelio.com"
authors = ["Nikho Sagala <nikhosagala@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6.8"
django = "^3.0.6"
beautifulsoup4 = "^4.9.0"
requests = "^2.23.0"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"
python-decouple = "^3.3"
behave = "^1.2.6"
behave-django = "^1.3.0"
invoke = "^1.4.1"
celery = "^4.4.2"
redis = "^3.5.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pre-commit = "^2.4.0"
flake8 = "^3.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
